requirements.txt
selenium
webdriver-manager
requests
beautifulsoup4
python-dotenv
tenacity
fake-useragent
sqlalchemy
tqdm
PyMuPDF
pandas
openpyxl